[SPLIT-4244] Some more refactor
