+++ Splitio-net-core-tests/Unit Tests/Cache/InMemorySimpleCacheTests.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Splitio.Domain;
-using Splitio.Services.Cache.Classes;
-using Splitio.Services.Shared;
+using Splitio.Services.Shared.Classes;
 
 namespace Splitio_Tests.Unit_Tests.Cache
 {
