+++ src/Splitio-net-core/Services/Client/Interfaces/ISplitClient.cs
@@ -14,7 +14,7 @@ namespace Splitio.Services.Client.Interfaces
         Dictionary<string, string> GetTreatments(string key, List<string> features, Dictionary<string, object> attributes = null);
 
         Dictionary<string, string> GetTreatments(Key key, List<string> features, Dictionary<string, object> attributes = null);
-        bool Track(string key, string trafficType, string eventType, double? value);
+        bool Track(string key, string trafficType, string eventType, double value);
         bool Track(string key, string trafficType, string eventType);
 
         void Destroy();
