+++ src/Splitio-net-core/Services/Shared/Interfaces/ISimpleCache.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Splitio.Services.Cache.Interfaces
+namespace Splitio.Services.Shared.Interfaces
 {
     public interface ISimpleCache<T>
     {
